+++ packages/react/src/components/participant/ParticipantTile.tsx
@@ -2,84 +2,23 @@ import * as React from 'react';
 import type { Participant, TrackPublication } from 'livekit-client';
 import { Track } from 'livekit-client';
 import type { ParticipantClickEvent, TrackReferenceOrPlaceholder } from '@livekit/components-core';
-import { isParticipantSourcePinned, setupParticipantTile } from '@livekit/components-core';
+import { isParticipantSourcePinned } from '@livekit/components-core';
 import { ConnectionQualityIndicator } from './ConnectionQualityIndicator';
 import { ParticipantName } from './ParticipantName';
 import { TrackMutedIndicator } from './TrackMutedIndicator';
 import {
-  useMaybeParticipantContext,
   ParticipantContext,
   useEnsureParticipant,
   useMaybeLayoutContext,
+  useMaybeParticipantContext,
   useMaybeTrackContext,
 } from '../../context';
-import { useFacingMode, useIsMuted, useIsSpeaking } from '../../hooks';
-import { mergeProps } from '../../utils';
 import { FocusToggle } from '../controls/FocusToggle';
 import { ParticipantPlaceholder } from '../../assets/images';
 import { ScreenShareIcon } from '../../assets/icons';
 import { VideoTrack } from './VideoTrack';
 import { AudioTrack } from './AudioTrack';
-
-/** @public */
-export interface ParticipantTileProps extends React.HTMLAttributes<HTMLDivElement> {
-  disableSpeakingIndicator?: boolean;
-  participant?: Participant;
-  source?: Track.Source;
-  publication?: TrackPublication;
-  onParticipantClick?: (event: ParticipantClickEvent) => void;
-}
-
-/** @public */
-export interface UseParticipantTileProps<T extends HTMLElement> extends React.HTMLAttributes<T> {
-  disableSpeakingIndicator?: boolean;
-  publication?: TrackPublication;
-  onParticipantClick?: (event: ParticipantClickEvent) => void;
-  htmlProps: React.HTMLAttributes<T>;
-  source: Track.Source;
-  participant: Participant;
-}
-
-/** @public */
-export function useParticipantTile<T extends HTMLElement>({
-  participant,
-  source,
-  publication,
-  onParticipantClick,
-  disableSpeakingIndicator,
-  htmlProps,
-}: UseParticipantTileProps<T>) {
-  const p = useEnsureParticipant(participant);
-  const mergedProps = React.useMemo(() => {
-    const { className } = setupParticipantTile();
-    return mergeProps(htmlProps, {
-      className,
-      onClick: (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-        // @ts-ignore
-        htmlProps.onClick?.(event);
-        if (typeof onParticipantClick === 'function') {
-          const track = publication ?? p.getTrack(source);
-          onParticipantClick({ participant: p, track });
-        }
-      },
-    });
-  }, [htmlProps, source, onParticipantClick, p, publication]);
-  const isVideoMuted = useIsMuted(Track.Source.Camera, { participant });
-  const isAudioMuted = useIsMuted(Track.Source.Microphone, { participant });
-  const isSpeaking = useIsSpeaking(participant);
-  const facingMode = useFacingMode({ participant, publication, source });
-  return {
-    elementProps: {
-      'data-lk-audio-muted': isAudioMuted,
-      'data-lk-video-muted': isVideoMuted,
-      'data-lk-speaking': disableSpeakingIndicator === true ? false : isSpeaking,
-      'data-lk-local-participant': participant.isLocal,
-      'data-lk-source': source,
-      'data-lk-facing-mode': facingMode,
-      ...mergedProps,
-    } as React.HTMLAttributes<HTMLDivElement>,
-  };
-}
+import { useParticipantTile } from '../../hooks';
 
 /** @public */
 export function ParticipantContextIfNeeded(
@@ -95,6 +34,15 @@ export function ParticipantContextIfNeeded(
   ) : (
     <>{props.children}</>
   );
+}
+
+/** @public */
+export interface ParticipantTileProps extends React.HTMLAttributes<HTMLDivElement> {
+  disableSpeakingIndicator?: boolean;
+  participant?: Participant;
+  source?: Track.Source;
+  publication?: TrackPublication;
+  onParticipantClick?: (event: ParticipantClickEvent) => void;
 }
 
 /**
