+++ packages/react/src/hooks/useParticipantTile.ts
@@ -1,0 +1,61 @@
+import type { ParticipantClickEvent } from '@livekit/components-core';
+import { setupParticipantTile } from '@livekit/components-core';
+import type { TrackPublication, Participant } from 'livekit-client';
+import { Track } from 'livekit-client';
+import React from 'react';
+import { useEnsureParticipant } from '../context';
+import { mergeProps } from '../mergeProps';
+import { useFacingMode } from './useFacingMode';
+import { useIsMuted } from './useIsMuted';
+import { useIsSpeaking } from './useIsSpeaking';
+
+/** @public */
+export interface UseParticipantTileProps<T extends HTMLElement> extends React.HTMLAttributes<T> {
+  disableSpeakingIndicator?: boolean;
+  publication?: TrackPublication;
+  onParticipantClick?: (event: ParticipantClickEvent) => void;
+  htmlProps: React.HTMLAttributes<T>;
+  source: Track.Source;
+  participant: Participant;
+}
+
+/** @public */
+export function useParticipantTile<T extends HTMLElement>({
+  participant,
+  source,
+  publication,
+  onParticipantClick,
+  disableSpeakingIndicator,
+  htmlProps,
+}: UseParticipantTileProps<T>) {
+  const p = useEnsureParticipant(participant);
+  const mergedProps = React.useMemo(() => {
+    const { className } = setupParticipantTile();
+    return mergeProps(htmlProps, {
+      className,
+      onClick: (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
+        // @ts-ignore
+        htmlProps.onClick?.(event);
+        if (typeof onParticipantClick === 'function') {
+          const track = publication ?? p.getTrack(source);
+          onParticipantClick({ participant: p, track });
+        }
+      },
+    });
+  }, [htmlProps, source, onParticipantClick, p, publication]);
+  const isVideoMuted = useIsMuted(Track.Source.Camera, { participant });
+  const isAudioMuted = useIsMuted(Track.Source.Microphone, { participant });
+  const isSpeaking = useIsSpeaking(participant);
+  const facingMode = useFacingMode({ participant, publication, source });
+  return {
+    elementProps: {
+      'data-lk-audio-muted': isAudioMuted,
+      'data-lk-video-muted': isVideoMuted,
+      'data-lk-speaking': disableSpeakingIndicator === true ? false : isSpeaking,
+      'data-lk-local-participant': participant.isLocal,
+      'data-lk-source': source,
+      'data-lk-facing-mode': facingMode,
+      ...mergedProps,
+    } as React.HTMLAttributes<T>,
+  };
+}
